requirements.txt
Flask==2.3.0
flask-cors==4.0.0
selenium==4.15.0
boto3==1.28.0
python-dotenv==1.0.0
requests==2.31.0
